Creation of the template for the packaging
I created a package for the project. The documentation will be used to
create the paper for the submission and the code will used through a
command line and a config file.

I updated the:

- AUTHORS.txt file
- HISTORY.txt file: restated the first bump
- setup.py file: updating of all required packages
- deletion of requirements.txt: everything is in setup.py now

All files aren't been tempered from the template generation and are
added as they were.

I moved the get_data.py file to the asc folder. It's at the moment not
called from the command line interface.

Signed-off-by: Matthieu Berjon <matthieu@berjon.net>