Adding of a cli subcommand to download data
I added a subcommand to download data. I changed the structure of of the
command line to create a new group. I added then 3 parameters for:

- the file containing the urls pointing to the data to download
- the temporary folder where to download the compressed file
- the data folder where to uncompress the files.

At the moment, the program check if the data folder exists and ask the
user if he wants to pursue if the folder exists.

Signed-off-by: Matthieu Berjon <matthieu@berjon.net>