Add a download method to the Data class
I added a `download()` method to the `Data` class. The method display a progress
bar according to the size of the file.
I removed the function that were used before and clean a bit the code.

I did a small correction as well to the `file_to_list()` method on the
line splitting. The function were returning a nested list, it now
returns a simple list.

Signed-off-by: Matthieu Berjon <matthieu@berjon.net>