Dependencies issues with python 3.x
I corrected an issue related to the installation of the package on
python 3.x. I had an issue with two libraries that are compatible python
2.7 only. I added one library to the `extra_requirement` setup with a
flag to specify the installation only on python 2.7 but forgot to remove
it from the requirement list. By doing that I discovered a second
library generating the same issue. I moved it as well to the right place
with the same flag. This should resolve completely the issue.

See also: #4

Signed-off-by: Matthieu Berjon <matthieu@berjon.net>