Update of Tox and Travis configuration for code coverage
I modified the configuration in order to compile the code coverage and
send the information to the coveralls service.

I modified the `.travis.yml` config file in order to install the package
coveralls. The file `tox.ini` has been modified in order to run the code
coverage according to the test suite.

See also: #5

Signed-off-by: Matthieu Berjon <matthieu@berjon.net>