Update of badges related to PyPI
I updated the badges links and details to corresponds to the exact
information given by the PyPI server as well as some external services
such as coveralls and pyup (that were already in places).

I put all of them in a table and separate them by categories to make it
a bit more clear I hope.

Signed-off-by: Matthieu Berjon <matthieu@berjon.net>