Table for badges in the README file
Update of badges related to PyPI

Update of badges related to PyPI

I updated the badges links and details to corresponds to the exact
information given by the PyPI server as well as some external services
such as coveralls and pyup (that were already in places).

I put all of them in a table and separate them by categories to make it
a bit more clear I hope.

Signed-off-by: Matthieu Berjon <matthieu@berjon.net>

Bug fix on badges

There were a small bug related to the badges. The table wasn't formatted
correctly. I added the right header for the table and indent the table
titles.

Signed-off-by: Matthieu Berjon <matthieu@berjon.net>

Bug fix badges table

There were still a display bug in the table badges. I decided to make
something simpler and wrote a simple table.

Signed-off-by: Matthieu Berjon <matthieu@berjon.net>