Bug fix: Update the config file parameters
I fixed the bug related to the update of the parameters. I moved it into
a proper function. There are still work to do on it such as ensuring the
casting of the data as strings before writing them.

Basically the `ConfigParser` class can only look for existing sections
or options if the content of the configuration file is loaded into
memory.

Resolves: #13

Signed-off-by: Matthieu Berjon <matthieu@berjon.net>