Added feature: Spectrogram calculation
I added as a set a functions the spectrogram calculation. At the moment,
the main function compute the spectrogram only of a sine signal
according to its frequency, sampling rate and length in sample. The
other function takes the output of spectrogram function to plot it.
The next step would be to compute the spectrogram on a real signal given
in a wav file.

See also: #10

Signed-off-by: Matthieu Berjon <matthieu@berjon.net>