Added feature: dynamic spectrogram computation
I added the computation of the dynamic spectrogram block by block.
In order to perform this operation, I added several functions:

- `spectrogram()` now computes the spectrogram for a real signal using
the 3rd parties library Scipy. The documentation has been updated
accordingly.
- `plot_spectrogram()` now can either display the figure or save it. See
the Todo part in the docstring to see the work that still need to be
done on this particular part.
- `audio_metadata()` extract some data from the audio file such as the
samplerate or channel number in order to be able to compute the
spectrogram on each channel properly. The name of the function isn't
really self explanatory and will need either to be changed or be worked
out to be divided in smaller functions.
- `process_audio()` is an empty function having in the docstring
information about the exact processing chain to respect to perform the
computation correclty.

Resolves: #10

Signed-off-by: Matthieu Berjon <matthieu@berjon.net>

Docstring fix

I fixed the docstrings to match the exact denomination regarding the
spectrogram computation. In the present code, we compute the dynamic
spectrogram and not the static one that needs a different computation.

See also: #10

Signed-off-by: Matthieu Berjon <matthieu@berjon.net>