Add a config sub command to the CLI
I changed the behaviour of the CLI by adding a `config` subcommand that
acts like the `config` subcommand in git. The CLI takes a parameter
(containing both the section and option of the INI format) and a value.
The program then either update the data if it already exists in the
configuration file or write it.

I added a utility function in order to extract the section and option
from the parameter given in the command. Because the config file is now
only used on this part of the CLI, I moved the object inside it. This
avoid a strange behaviour. If the configparser object is declared of the
top of the file, when we pass the pointer of this object to a function,
its type is then transformed as a click.core.Command type. I didn't find
any explanation why this behaviour happens.

Resolves: #24

Signed-off-by: Matthieu Berjon <matthieu@berjon.net>