Configuration file documentation
I created a complete new part related to the different option that can
be used in the configuration file in order to give the user an idea
about what can be configured easily without having to read the source
code. I added this file into the documentation index.

See also: #24

Signed-off-by: Matthieu Berjon <matthieu@berjon.net>