Bug fix: test suite failure on python 3.4
It seems that the build systematically fails on python 3.4 on Travi-CI.
It seems that Tkinker is missing for some reasons even if it's not the
front end used by matplotlib. In order to fix that, before running the
tests, I update the package manager (Ubuntu) and then install
`python3-tk` package. This should resolve the issue.

See also: #29

Signed-off-by: Matthieu Berjon <matthieu@berjon.net>