Check database samplerate sanity
I created a small script that check the sanity of the audio database
regarding the samplerate. It looks for the configuration samplerate, the
data path and extract all wav files and confront the data. It outputs
the result using a logging library.

See also: #21

Signed-off-by: Matthieu Berjon <matthieu@berjon.net>

Update of the script docstring

I updating the docstring with a small todo in order to not forget some
stuff to do that are important.

See also: #21

Signed-off-by: Matthieu Berjon <matthieu@berjon.net>