Bug fix: test failure because configuration file is empty
Because at the installation the configuration file is empty, one of the
tests fails because it's impossible to find the correct data. At the
moment, I decided to set the correct data in the instanciated object.
This fix the issue but probably mocking would be a better option.

See also: #31

Signed-off-by: Matthieu Berjon <matthieu@berjon.net>