Bug fix: compilation issue on Fedora 26 and python 2.7
I had an issue at the installation on Fedora 26 and python 2.7. It
appears that the issue is related to a missing package called
`redhat-rpm-config`. Once installed, the compilation process is a
success.

In order to avoid any future issue for the user, I added a FAQ in the
documentation helping the resolution.

Resolves: #34

Signed-off-by: Matthieu Berjon <matthieu@berjon.net>