Switch to librosa lib
I did two main actions in this commit. I chose to move everything
related to audio processing in the `audio.py` file and delete `asc.py`
that hadn't any real meaning here.

I switched from scipy to librosa library to compute the dynamic
spectrograms. I did so because it's a recommended library in the
research paper and because it will make all the audio processing easier.

I did several changes listed here:

- I renamed extract_audio_metadata into process_audio which is more
accurate and because the metadata are now obtained in another function.
- I wrote the small function extract_audio_data which look for the audio
file samplerate and number of channel. Some other data could added later
if necessary (such as the depth of the data and so on)
- I modified the spectrogram function to match the use of the librosa
lib.
- I deleted the `plot_spectrogram()` function because of the
architecture of librosa which now is available directly inside the
`spectrogram()` function and enabled through a flag.

Some improvements can be made such as:

- Create an `AudioData` class that will have data structure adapted to
handle audio data and metadata a bit more easily. The plot saving isn't
enabled yet. It's necessary to grab the folder where to save them into
the config file and find a good way to name the files. It would be
interesting to save the plots without any legend, margin or padding in
the pictures.

Resolves: #33

Signed-off-by: Matthieu Berjon <matthieu@berjon.net>