Adding of a note
I added a note regarding the frame size and hop size. This information
comes from the base line described by the reference implementation.

It's not a mandatory ensemble of parameters but it's still interesting
to keep it close.

The frame size is important and has a huge impact on the spectrogram
computation.

Signed-off-by: Matthieu Berjon <matthieu@berjon.net>