Bug fixes: frame size conversion, mono file processing
This patch provides two bug fixes. The first one relates to the
processing of mono files. It wasn't working because 1D numpy arrays
can't be called like in Matlab/Octave or nympy 2D arrays.
In order to fix that, I check if the number of columns is inferior to
two. If it's the case, I load the 1D array otherwise I do the process
for each channel.

The second bug fix was a wrong unit definition. The frame size wasn't
expressed in frames but in samples leading to confusion. Now the frame
size is fixed and 1 frame corresponds to 512 samples. At the moment the
conversion rate can't be changed and corresponds to the function
provided by `librosa`

Resolves: #36, #40
See also: #12

Signed-off-by: Matthieu Berjon <matthieu@berjon.net>