Static spectrogram computation
A minor changed has been done. the `spectrogram()` function has been
renamed into `dynamic_spectrogram()` in order to fit the paper
description and be more accurate in the audio processing workflow.

Then the  `static_spectrogram()` function has been added. This has an
impact of `process_audio()` from which the function is called. I added
the necessary arguments to pas such the the number of bands for the mel
decomposition of the signal and the max frequency. I added as well the
display argument for future use in debug and logging mode.

Finally I modified the cli in order to support those arguments.

Resolves: #41
See also: #10

Signed-off-by: Matthieu Berjon <matthieu@berjon.net>