Bug fix: Unzip error after the data download
The unzipping of the archive was returning an error when the object was
calling the `unzip()` method. In fact the code base has been modified
and it's now called `unzip_data()`. I renamed the function and tested
it. It's working.

I took as well the opportunity of fixing few syntax issues such as
trailing spaces to enforce the PEP8 compliance.

There is still once suspending issue when the unzipping is exectuted.
The data are separated in several folders which seems quite annoying at
first and maybe need to be addressed.

Resolves: #43
See also: #44

Signed-off-by: Matthieu Berjon <matthieu@berjon.net>