Installation bug on Travis
It seems that Pip dependencies installation order creates some issues by
installation numba first when it depends on Numpy leading to an error.
I put numpy at the top of the list of requirement libs hoping that it
would correct this issue. As far as I read on the subject, I don't think
it would change anything but it still worths a try.

Signed-off-by: Matthieu Berjon <matthieu@berjon.net>