No travis-ci test on python 3.x
At the moment, I have too many issues with travis related to python 3.x.

For some reasons, on the python 3.4 image, I have an issue with numba
that requires llvm 4.x. On Fedora I don't this issue whatsoever and on
Ubuntu Trusty, it requires to install too many useless things and still
do not work out. It seems related to the Travis docker image and I lost
too much time on this issue and can't solve it.

On the python 3.5 docker image, there is a unicode issue on the pyenv.
According to what I read, this is a bug on Travis side that isn't
resolved yet.

Signed-off-by: Matthieu Berjon <matthieu@berjon.net>