Update of read/write config, add of user prompt
I updated the read/write configuration functions. Basically I moved
everything inside such as the call of the configparser object and the
opening of the configuration file. The only thing that I call now is the
path of the configuration file.

I added as well two functions for reading an input user. First a
specific function to get a generic input and another one returning a
boolean for true or false question.

Signed-off-by: Matthieu Berjon <matthieu@berjon.net>