Add samples to the tree
I added few samples regarding the configuration file to give some ideas
to the user of the syntax. They can refer then to the documentation for
all the existing options.

I added as well the url_file list in order to save some time and avoid
the user to write himself this file that is cumbersome.

Signed-off-by: Matthieu Berjon <matthieu@berjon.net>