Add of a function to create filename pattern
I wrote a small function that takes a path and an extension as mandatory
arguments and any other arguments to create a file name pattern.

This is used mainly to save the spectrograms in the right folder
following a specific naming pattern.

I updated as well the write/read config functions that are now
simplified and take only the section/option/data tuple and nothing more
which is more convenient.
I adapted the test suite accordingly. I still need to update the
docstrings.

Signed-off-by: Matthieu Berjon <matthieu@berjon.net>