Adding of a function
In order to be able to put all the data in a unique folder I wrote a
small function that move files into a specified folder.

The function can take as arguments a specific pattern for the files to
move, the folder where are located the files and the folder destination.

Signed-off-by: Matthieu Berjon <matthieu@berjon.net>