Path change for docstring documentation
I decided to move the docstring documentation generated files from the
`source` folder into the root documentation folder. This seems more
coherent with the current setup and keep all the RST files into one
unique folder. I removed as well the `.gitignore` because its solely
purpose was to avoid that but it's relevant anymore. Moreover not
pushing them on the tree would cause the documentation generation on
ReadTheDoc to not compile properly.

Signed-off-by: Matthieu Berjon <matthieu@berjon.net>