Force default value overriding
I enforce the use of the config file over the default values if the user
chooses to change them by using the `asc config` command line or
directly writing them in the `~/.ascrc` dotfile.

I fixed as well a small bug by forcing the config type into a
`click.STRING`. This casting avoids the use of  simple quotes in the
command line.

Resolves: #48

Signed-off-by: Matthieu Berjon <matthieu@berjon.net>