Move the unzipped data to the data folder
I fixed the function that is supposed to move all audio files into the
data folder. This relates to the archive internal architecture that was
supposed to be complicated but in fact all archives unzip their data
into the same folder. There are no architecture complexity whatsoever,
just need to move the `audio` folder into the data root path.

Resolves: #44

Signed-off-by: Matthieu Berjon <matthieu@berjon.net>

Update of the docstring

I updated the docstring in order to keep track of the arguments and have
a proper description of the function use.

Signed-off-by: Matthieu Berjon <matthieu@berjon.net>