Create the subfolders whenever the setup command is called
There were an issue regarding the creation of subfolders. If a user
decided to not start its project using the `setup` command but decided
to run it afterwards, the subfolder weren't created because if the root
folder was already existing we were supposing that the subfolders were
existing too. I corrected that by checking each subfolder.

Resolves: #50

Signed-off-by: Matthieu Berjon <matthieu@berjon.net>