Update of documentation and related docstring
I updated the `setup` docstring to have a proper documentation and help
into the command line system. I also updated the documentation related
to the usage of both the library and command line.

Signed-off-by: Matthieu Berjon <matthieu@berjon.net>

Update of the docstring

I updated the docstring in the `setup` command in order to have an
explanation about what the command does and have a helper in the CLI.

Signed-off-by: Matthieu Berjon <matthieu@berjon.net>