Adding of an entry in the FAQ
I added an entry regarding a potential issue that might appear for users
trying to launch the program on a remote server without any XWindow
available. I foung the workaround and explained it.

I as well make a few precision on the installation issue regarding the
missing of GCC on Python 2.7 and give the right package to install on
Debian related distributions.

I fixed as well a syntax issue in the RST file by itself.

Resolves: #55

Signed-off-by: Matthieu Berjon <matthieu@berjon.net>