Bug fix: unzipped audio data in the wrong folder
There were a bug regarding the audio data unzipping. They were moved
into the wrong folder. There were 2 issues:

- the default `data` folder in the config.py file didn't match the
  default value in the CLI setup
- In the CLI command to download and unzip the data in the audio folder,
  the path extraction from the `~/.ascrc` file was unvalid.

The first issue was corrected by giving to the `config.py` file the
right path `asc-data`.

Regarding the second issue, it was trying to extract a `data` path value
from the `~/.ascrc` config file that doesn't exist. I fixed it by
calling the right value which is `audio` in the `path` section of the
config file.

Resolves: #54

Signed-off-by: Matthieu Berjon <matthieu@berjon.net>